Fixed CV override for GOVSS and SwimScore Copy paste error, cv is a double not an int as cp

+++ src/Metrics/SwimScore.cpp
@@ -225,7 +225,7 @@ class STP : public RideMetric {
         int zoneRange = item->paceZoneRange;
 
         // did user override for this ride?
-        double cv = item->getText("CV","0").toInt();
+        double cv = item->getText("CV","0").toDouble();
 
         // not overriden so use the set value
         // if it has been set at all
